+++ server/routes/comments.js
@@ -7,17 +7,19 @@ const auth = require('../controllers/auth');
 
 const { validateCommentContent } = require('../controllers/validation');
 
+const can = require('../permissions/comments')
+
 const prefix = '/api/v1/comments';
 const router = new Router({ prefix: prefix });
 
-router.get('/:post_id', getAllComments);
+router.get('/:post_id', getAllCommentsOfAPost);
 router.post('/:post_id', bodyParser(), auth.requireJWT, validateCommentContent, createNewComment);
 router.put('/:comment_id', bodyParser(), auth.requireJWT, validateCommentContent, editComment);
 router.del('/:comment_id', auth.requireJWT, deleteComment);
 
-async function getAllComments(ctx) {
+async function getAllCommentsOfAPost(ctx) {
     const postID = ctx.params.post_id;
-    ctx.body = await model.getAllComments(postID);
+    ctx.body = await model.getAllCommentsOfAPost(postID);
 }
 
 async function createNewComment(ctx) {
@@ -44,8 +46,6 @@ async function createNewComment(ctx) {
     }
 }
 
-// TODO: check if the user is allow to update the comment
-// if the post is user's comment || if the user's role is admin ? 
 async function editComment(ctx) {
     const body = ctx.request.body;
     if (!body.content) {
@@ -55,11 +55,30 @@ async function editComment(ctx) {
     }
 
     const commentID = ctx.params.comment_id;
+    const comment = await model.getCommentByCommentID(commentID);
+
+    if (!comment) {
+        ctx.status = 404;
+        ctx.body = { error: "Comment not found." }; 
+        return;
+    }
+
+    const permission = can.update(ctx.state.user, comment);
+
+    if (!permission.granted) {
+        ctx.status = 403;
+        ctx.body = { error: "You do not own this comment." };
+        return;
+    }
+
     try {
         const result = await model.updateComment(commentID, body.content);
         if (result.affectedRows) {
             ctx.status = 201;
             ctx.body = { message: `Comment edited for ${commentID}` };
+        } else {
+            ctx.status = 400;
+            ctx.body = { error: "Post could not be updated." };
         }
 
     } catch (err) {
@@ -69,18 +88,36 @@ async function editComment(ctx) {
     }
 }
 
-// TODO: check if the user is allow to delete the comment
-// if the post is user's post || if the comment is user's comment || if the user's role is admin ? 
+// TODO: let the post owner delete any comment on his post
 async function deleteComment(ctx) {
     const commentID = ctx.params.comment_id;
+
+    const comment = await model.getCommentByCommentID(commentID);
+    if (!comment) {
+        ctx.status = 404;
+        ctx.body = { error: "Comment not found."};
+        return;
+    }
+
+    const permission = can.delete(ctx.state.user, comment);
+
+    if (!permission.granted) {
+        ctx.status = 403;
+        ctx.body = { error: "You do not own this comment." };
+        return;
+    }
+
     try {
         const result = await model.deleteComment(commentID);
         if (result.affectedRows) {
             ctx.status = 200;
             ctx.body = { message: "Comment deleted." };
+        } else {
+            ctx.status = 400;
+            ctx.body = { error: "Comment not found." };
         }
     } catch (err) {
-        console.error();
+        console.error("Error deleting comment: ", err);
         ctx.status = 500;
         ctx.body = { error: "Internal Server Error." }
     }
