+++ server/routes/users.js
@@ -7,13 +7,16 @@ const auth = require('../controllers/auth');
 const jwt = require('jsonwebtoken');
 const { JWT_SECRETKEY } = require('../config');
 
-const { validateUserRegistration, validateUserLogin } = require('../controllers/validation');
+const { validateUserRegistration, validateUserLogin, validateUsernameUpdate } = require('../controllers/validation');
+
+const can = require('../permissions/users');
 
 const prefix = '/api/v1/users';
 const router = new Router({ prefix: prefix });
 
 router.post('/', bodyParser(), validateUserRegistration, registerUser);
 router.post('/login', bodyParser(), validateUserLogin, auth.requireLocal, loginUser);
+router.put('/', bodyParser(), auth.requireJWT, validateUsernameUpdate, changeUserName);
 router.del('/:id', auth.requireJWT, deleteUser);
 
 async function registerUser(ctx) {
@@ -57,6 +60,7 @@ async function loginUser(ctx) {
     const payload = {
         ID: user.id,
         username: user.username,
+        email: user.email,
         role: user.role
     };
 
@@ -70,10 +74,67 @@ async function loginUser(ctx) {
     };
 }
 
-// TODO: check if the user is allow to delete the account
-// if the account is user's account || if the user's role is admin ? 
+async function changeUserName(ctx) {
+    const body = ctx.request.body;
+
+    if (!body.username) {
+        ctx.status = 400;
+        ctx.body = { error: "New username required to update post." };
+        return;
+    }
+
+    const userID = ctx.state.user.id;
+    const user = await model.getUserByID(userID);
+
+    if (!user) {
+        //This should never happen since we are getting user from ctx.state.id
+        ctx.status = 404;
+        ctx.body = { error: "User not found" };
+        return;
+    }
+
+    const permission = can.update(ctx.state.user, user);
+
+    if (!permission.granted) {
+        ctx.status = 403; 
+        ctx.body = { error: "You do not own this account." };
+        return;
+    }
+    try {
+        const result = await model.updateUserName(userID, body.username);
+        if (result.affectedRows) {
+            ctx.status = 200;
+            ctx.body = { message: `Username for id ${userID} is changed.` };
+        } else {
+            ctx.status = 400;
+            ctx.body = { error: "Username could not be updated." };
+        }
+
+    } catch (err) {
+        console.error("Error occured while updating username.", err);
+        ctx.status = 500;
+        ctx.body = { error: "Internal Server Error." };
+    }
+}
+
 async function deleteUser(ctx) {
     const userID = ctx.params.id;
+
+    const user = await model.getUserByID(userID);
+    if (!user) {
+        ctx.status = 404;
+        ctx.body = { error: "User not found." };
+        return; 
+    }
+
+    const permission = can.delete(ctx.state.user, user);
+    
+    if (!permission.granted) {
+        ctx.status = 403; 
+        ctx.body = { error: "You do not own this account." };
+        return;
+    }
+
     try {
         const result = await model.deleteUser(userID);
         if (result.affectedRows) {
